move helper methods for fixtures into comon test.utils.py
